pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "showup_squared_generator"
version = "1.0.0"
requires-python = ">=3.8"
dependencies = [
    "anthropic",
    "claude_api",
    "loguru",
    "numpy",
    "pandas",
    "python-dotenv",
    "openai",
    "python-docx",
    "Pillow",
    "markdown",
    "aiohttp",
    "azure-cognitiveservices-speech",
    "scikit-learn",
    "sentence-transformers",
    "setuptools",
    "tqdm"
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "isort",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests", "dev_docs"]
